Migrate DataForTable and DateForPartition
